pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Mani <mani.mozaffar@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.96.0"
asyncpg = "^0.27.0"
sqlalchemy = "^2.0.16"
uvicorn = "^0.22.0"
pydantic = {extras = ["dotenv"], version = "^1.10.9"}
aiosqlite = "^0.19.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
isort = "^5.12.0"
opentelemetry-sdk = "^1.18.0"
bcrypt = "^4.0.1"
alembic = "^1.11.1"
rich = "^13.4.2"
redis = "^4.5.5"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.0"
greenlet = "^2.0.2"
pyotp = "^2.8.0"
qrcode = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
